Renamed OneSignalHelpers file
